requirements.txt
python-dotenv==1.0.1
langchain==0.2.16
langchain-community==0.2.16
sentence-transformers==3.0.1
langchain-huggingface==0.0.3
chromadb==0.5.5
langchain-chroma==0.1.2
langchain-google-genai==1.0.10
unstructured==0.15.9
unstructured[md]
unstructured[xlsx]
streamlit
unstructured[pdf]